model/requirements.txt
pandas~=2.1.4
numpy~=1.26.3
tensorflow~=2.15.0
pillow~=10.2.0